pyproject.toml
[tool.poetry]
name = "habit-hive"
version = "0.1.0"
description = "habit reminder app"
authors = ["Natalia D"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.7"
psycopg2 = "^2.9.9"
ipython = "^8.21.0"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
flake8 = "^7.0.0"
coverage = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
